Refactor: extract Configuration and ModifiableConfiguration interfaces
