+++ src/main/java/de/fakeller/performance/variability/configuration/Configuration.java
@@ -1,85 +1,24 @@
 package de.fakeller.performance.variability.configuration;
 
-import de.fakeller.performance.variability.feature.BaseFeatureModel;
-import de.fakeller.performance.variability.feature.UnknownFeatureException;
+import de.fakeller.performance.variability.feature.FeatureModel;
 
-import java.util.*;
+import java.util.List;
 
 /**
- * Defines a configuration of the {@link BaseFeatureModel}, such that each {@link FEATURE} is either enabled or disabled.
+ * Defines a configuration of the {@link FeatureModel}, such that each {@link FEATURE} is either enabled or disabled.
+ *
+ * @param <FEATURE> the backing class representing a feature. Features must properly implement {@link Object#hashCode()}
+ *                  and {@link Object#equals(Object)}.
  */
-public class Configuration<FEATURE> {
-
-    private final BaseFeatureModel<FEATURE> fm;
-
-    private final Map<FEATURE, Boolean> isEnabled = new LinkedHashMap<>();
-
-    public Configuration(final BaseFeatureModel<FEATURE> fm) {
-        this.fm = fm;
-        this.fm.getFeatures().forEach(f -> this.isEnabled.put(f, false));
-    }
-
-
-    public Configuration<FEATURE> enable(final FEATURE... features) {
-        return this.enable(Arrays.asList(features));
-    }
-
-    public Configuration<FEATURE> enable(final Collection<FEATURE> features) {
-        this.setConfiguration(features, true);
-        return this;
-    }
-
-    /**
-     * Sets all feature flags to enabled.
-     */
-    public Configuration<FEATURE> enableAll() {
-        return this.enable(this.isEnabled.keySet());
-    }
-
-
-    public Configuration<FEATURE> disable(final FEATURE... features) {
-        return this.disable(Arrays.asList(features));
-    }
-
-    public Configuration<FEATURE> disable(final Collection<FEATURE> features) {
-        this.setConfiguration(features, false);
-        return this;
-    }
-
-    /**
-     * Sets all feature flags to enabled.
-     */
-    public Configuration<FEATURE> disableAll() {
-        return this.disable(this.isEnabled.keySet());
-    }
+public interface Configuration<FEATURE> {
 
     /**
      * Determines whether the given feature is enabled or not.
      */
-    public boolean isEnabled(final FEATURE feature) {
-        this.assertContainsFeature(feature);
-        return this.isEnabled.get(feature);
-    }
+    boolean isEnabled(FEATURE feature);
 
     /**
      * Returns an ordered list of boolean flags indicating whether the feature at index i is enabled or not.
      */
-    public List<Boolean> getFeatureFlags() {
-        return new ArrayList<>(this.isEnabled.values());
-    }
-
-
-    private void setConfiguration(final Collection<FEATURE> features, final boolean enabled) {
-        features.forEach(feature -> {
-            this.assertContainsFeature(feature);
-            this.isEnabled.put(feature, enabled);
-        });
-    }
-
-    private void assertContainsFeature(final FEATURE feature) {
-        if (!this.fm.hasFeature(feature)) {
-            throw new UnknownFeatureException(String.format("The feature model %s does not have feature %s. You must specify all valid features when constructing the feature model!", this.fm, feature));
-        }
-    }
-
+    List<Boolean> getFeatureFlags();
 }
