+++ src/test/java/de/fakeller/performance/variability/analysis/BaseContextTest.java
@@ -4,7 +4,7 @@ import de.fakeller.performance.analysis.AnalysisContext;
 import de.fakeller.performance.analysis.PerformanceAnalyzer;
 import de.fakeller.performance.analysis.result.PerformanceResult;
 import de.fakeller.performance.variability.SystemProvider;
-import de.fakeller.performance.variability.configuration.Configuration;
+import de.fakeller.performance.variability.configuration.BaseConfiguration;
 import de.fakeller.performance.variability.configuration.ConfigurationProvider;
 import de.fakeller.performance.variability.feature.BaseFeatureModel;
 import org.junit.Before;
@@ -47,7 +47,7 @@ public class BaseContextTest {
 
     @Test
     public void analyze_withNoConfigurations() throws Exception {
-        when(this.configurationProvider.configurations()).thenReturn(new ArrayList<Configuration<String>>().iterator());
+        when(this.configurationProvider.configurations()).thenReturn(new ArrayList<BaseConfiguration<String>>().iterator());
         this.sut.analyze();
         assertNotNull(this.sut.getResults());
     }
@@ -55,8 +55,8 @@ public class BaseContextTest {
     @Test
     public void analyze_withSomeConfigurations() throws Exception {
         when(this.configurationProvider.configurations()).thenReturn(Stream.of(
-                new Configuration<>(this.fm),
-                new Configuration<>(this.fm).enable(this.fm.get(0)).enable(this.fm.get(2))
+                new BaseConfiguration<>(this.fm),
+                new BaseConfiguration<>(this.fm).enable(this.fm.get(0)).enable(this.fm.get(2))
         ).iterator());
         when(this.systemProvider.systemFor(any())).thenReturn("x");
         final AnalysisContext ac = mock(AnalysisContext.class);
