+++ src/test/java/de/fakeller/performance/variability/configuration/BaseConfigurationTest.java
@@ -12,7 +12,7 @@ import static org.junit.Assert.*;
 /**
  * @author Fabian Keller
  */
-public class ConfigurationTest {
+public class BaseConfigurationTest {
 
     private BaseFeatureModel<String> fm;
 
@@ -23,7 +23,7 @@ public class ConfigurationTest {
 
     @Test
     public void enable_singleFeature() throws Exception {
-        final Configuration<String> cfg = new Configuration<>(this.fm);
+        final BaseConfiguration<String> cfg = new BaseConfiguration<>(this.fm);
         assertFalse(cfg.isEnabled("A"));
         assertFalse(cfg.isEnabled("C"));
         cfg.enable("A");
@@ -36,7 +36,7 @@ public class ConfigurationTest {
 
     @Test
     public void enable_collectionOfFeatures() throws Exception {
-        final Configuration<String> cfg = new Configuration<>(this.fm);
+        final BaseConfiguration<String> cfg = new BaseConfiguration<>(this.fm);
         assertFalse(cfg.isEnabled("A"));
         assertFalse(cfg.isEnabled("C"));
         cfg.enable(Arrays.asList("A", "C"));
@@ -46,13 +46,13 @@ public class ConfigurationTest {
 
     @Test(expected = UnknownFeatureException.class)
     public void enable_withUnknownFeature_throwsException() throws Exception {
-        final Configuration<String> cfg = new Configuration<>(this.fm);
+        final BaseConfiguration<String> cfg = new BaseConfiguration<>(this.fm);
         cfg.enable("undefined_feature");
     }
 
     @Test
     public void disable_singleFeature() throws Exception {
-        final Configuration<String> cfg = new Configuration<>(this.fm).enableAll();
+        final BaseConfiguration<String> cfg = new BaseConfiguration<>(this.fm).enableAll();
         assertTrue(cfg.isEnabled("A"));
         assertTrue(cfg.isEnabled("C"));
         cfg.disable("A");
@@ -65,7 +65,7 @@ public class ConfigurationTest {
 
     @Test
     public void disable_collectionOfFeatures() throws Exception {
-        final Configuration<String> cfg = new Configuration<>(this.fm).enableAll();
+        final BaseConfiguration<String> cfg = new BaseConfiguration<>(this.fm).enableAll();
         assertTrue(cfg.isEnabled("A"));
         assertTrue(cfg.isEnabled("C"));
         cfg.disable(Arrays.asList("A", "C"));
@@ -75,13 +75,13 @@ public class ConfigurationTest {
 
     @Test(expected = UnknownFeatureException.class)
     public void disable_withUnknownFeature_throwsException() throws Exception {
-        final Configuration<String> cfg = new Configuration<>(this.fm);
+        final BaseConfiguration<String> cfg = new BaseConfiguration<>(this.fm);
         cfg.disable("undefined_feature");
     }
 
     @Test
     public void isEnabled_byDefault_isFalse() throws Exception {
-        final Configuration<String> cfg = new Configuration<>(this.fm);
+        final BaseConfiguration<String> cfg = new BaseConfiguration<>(this.fm);
         assertFalse(cfg.isEnabled("A"));
         assertFalse(cfg.isEnabled("B"));
         assertFalse(cfg.isEnabled("C"));
@@ -90,13 +90,13 @@ public class ConfigurationTest {
 
     @Test(expected = UnknownFeatureException.class)
     public void isEnabled_withUnknownFeature_throwsException() throws Exception {
-        final Configuration<String> cfg = new Configuration<>(this.fm);
+        final BaseConfiguration<String> cfg = new BaseConfiguration<>(this.fm);
         cfg.isEnabled("undefined_feature");
     }
 
     @Test
     public void getFeatureFlags() throws Exception {
-        final Configuration<String> cfg = new Configuration<>(this.fm);
+        final BaseConfiguration<String> cfg = new BaseConfiguration<>(this.fm);
         assertEquals(Arrays.asList(false, false, false, false), cfg.getFeatureFlags());
 
         cfg.enable("B");
